+++ lib/widgets/recommendation_widget.dart
@@ -6,16 +6,16 @@ import 'package:critic/blocs/other_profile/other_profile_bloc.dart'
     as OTHER_PROFILE_BP;
 import 'package:critic/models/recommendation_model.dart';
 import 'package:critic/services/modal_service.dart';
-import 'package:expansion_card/expansion_card.dart';
+import 'package:expansion_tile_card/expansion_tile_card.dart';
 import 'package:flutter/material.dart';
 import 'package:timeago/timeago.dart' as timeago;
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 class RecommendationWidget extends StatefulWidget {
   const RecommendationWidget({
-    Key key,
-    @required this.recommendation,
-    @required this.delete,
+    Key? key,
+    required this.recommendation,
+    required this.delete,
   }) : super(key: key);
 
   final RecommendationModel recommendation;
@@ -26,7 +26,7 @@ class RecommendationWidget extends StatefulWidget {
 }
 
 class _RecommendationWidgetState extends State<RecommendationWidget> {
-  int _critiqueMessageCharCount = 75;
+  // int _critiqueMessageCharCount = 75;
 
   @override
   void initState() {
@@ -37,7 +37,7 @@ class _RecommendationWidgetState extends State<RecommendationWidget> {
   Widget build(BuildContext context) {
     return Padding(
       padding: EdgeInsets.all(10),
-      child: ExpansionCard(
+      child: ExpansionTileCard(
         leading: InkWell(
           onTap: () {
             Route route = MaterialPageRoute(
@@ -54,7 +54,7 @@ class _RecommendationWidgetState extends State<RecommendationWidget> {
             Navigator.push(context, route);
           },
           child: CachedNetworkImage(
-            imageUrl: '${widget.recommendation.sender.imgUrl}',
+            imageUrl: '${widget.recommendation.sender!.imgUrl}',
             imageBuilder: (context, imageProvider) => CircleAvatar(
               backgroundImage: imageProvider,
             ),
@@ -62,32 +62,34 @@ class _RecommendationWidgetState extends State<RecommendationWidget> {
             errorWidget: (context, url, error) => Icon(Icons.error),
           ),
         ),
-        borderRadius: 20,
-        background: ColorFiltered(
-          colorFilter: ColorFilter.mode(
-            Colors.black.withOpacity(0.7),
-            BlendMode.darken,
-          ),
-          child: Image.network(
-            '${widget.recommendation.movie.poster}',
-            fit: BoxFit.cover,
-            height: 200,
-            width: double.infinity,
-          ),
+        borderRadius: BorderRadius.all(
+          Radius.circular(20),
         ),
+        // background: ColorFiltered(
+        //   colorFilter: ColorFilter.mode(
+        //     Colors.black.withOpacity(0.7),
+        //     BlendMode.darken,
+        //   ),
+        //   child: Image.network(
+        //     '${widget.recommendation.movie!.poster}',
+        //     fit: BoxFit.cover,
+        //     height: 200,
+        //     width: double.infinity,
+        //   ),
+        // ),
         title: Container(
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: <Widget>[
               Text(
-                '${widget.recommendation.movie.title}',
+                '${widget.recommendation.movie!.title}',
                 style: TextStyle(
                     color: Colors.white,
                     fontSize: 16,
                     fontWeight: FontWeight.bold),
               ),
               Text(
-                  '${widget.recommendation.sender.username}, ${timeago.format(widget.recommendation.created, allowFromNow: true)}',
+                  '${widget.recommendation.sender!.username}, ${timeago.format(widget.recommendation.created, allowFromNow: true)}',
                   style: TextStyle(color: Colors.white, fontSize: 14)),
             ],
           ),
@@ -110,19 +112,20 @@ class _RecommendationWidgetState extends State<RecommendationWidget> {
             child: Row(
               mainAxisAlignment: MainAxisAlignment.spaceAround,
               children: [
-                RaisedButton(
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                  ),
-                  child: Row(
-                    children: [
-                      Icon(Icons.movie),
-                      SizedBox(
-                        width: 10,
+                ElevatedButton(
+                  style: ButtonStyle(
+                    backgroundColor: MaterialStateProperty.all(
+                        Theme.of(context).canvasColor),
+                    textStyle: MaterialStateProperty.all(
+                      TextStyle(color: Colors.white),
+                    ),
+                    shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+                      RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(8.0),
                       ),
-                      Text('Go To Movie'),
-                    ],
+                    ),
                   ),
+                  child: Text('Go To Movie'),
                   onPressed: () {
                     Route route = MaterialPageRoute(
                       builder: (context) => BlocProvider(
@@ -138,33 +141,32 @@ class _RecommendationWidgetState extends State<RecommendationWidget> {
 
                     Navigator.push(context, route);
                   },
-                  color: Theme.of(context).buttonColor,
-                  textColor: Colors.white,
                 ),
-                RaisedButton(
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                  ),
-                  child: Row(children: [
-                    Icon(Icons.delete),
-                    SizedBox(
-                      width: 10,
+                ElevatedButton(
+                  style: ButtonStyle(
+                    backgroundColor: MaterialStateProperty.all(
+                        Theme.of(context).canvasColor),
+                    textStyle: MaterialStateProperty.all(
+                      TextStyle(color: Colors.white),
                     ),
-                    Text('Delete')
-                  ]),
+                    shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+                      RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(8.0),
+                      ),
+                    ),
+                  ),
+                  child: Text('Delete'),
                   onPressed: () async {
-                    final bool confirm = await locator<ModalService>()
+                    final bool? confirm = await locator<ModalService>()
                         .showConfirmation(
                             context: context,
                             title: 'Delete Recommendation',
                             message: 'Are you sure?');
 
-                    if (!confirm) return;
+                    if (confirm == null || !confirm) return;
                     widget.delete();
                   },
-                  color: Theme.of(context).buttonColor,
-                  textColor: Colors.white,
-                )
+                ),
               ],
             ),
           )
